requirements.txt
django
djangorestframework
django-cors-headers
dj-database-url
python-dotenv
python-jose
gunicorn
psycopg2-binary
whitenoise